[#48991] Trigger synchronization on projects_storage mode changes.
https://community.openproject.org/work_packages/48991
