+++ modules/storages/app/services/storages/project_storages/create_service.rb
@@ -36,6 +36,7 @@ module Storages::ProjectStorages
 
       project_storage = service_call.result
       add_historical_data(service_call) if project_storage.project_folder_mode.to_sym != :inactive
+      Helper.trigger_nextcloud_synchronization(project_storage.project_folder_mode)
 
       service_call
     end
@@ -45,7 +46,7 @@ module Storages::ProjectStorages
     def add_historical_data(service_call)
       project_storage = service_call.result
       last_project_folder_result =
-        LastProjectFolderPersistenceHelper.create_last_project_folder(
+        Helper.create_last_project_folder(
           user:,
           projects_storage_id: project_storage.id,
           origin_folder_id: project_storage.project_folder_id,
