+++ modules/storages/app/services/storages/project_storages/update_service.rb
@@ -37,6 +37,7 @@ module Storages::ProjectStorages
 
       project_storage = service_call.result
       add_historical_data(service_call) if project_storage.project_folder_mode.to_sym != :inactive
+      Helper.trigger_nextcloud_synchronization(model.project_folder_mode)
 
       service_call
     end
@@ -53,14 +54,14 @@ module Storages::ProjectStorages
 
       last_project_folder_result =
         if project_folder.nil?
-          LastProjectFolderPersistenceHelper.create_last_project_folder(
+          Helper.create_last_project_folder(
             user:,
             projects_storage_id: project_storage.id,
             origin_folder_id: project_storage.project_folder_id,
             mode: project_storage.project_folder_mode
           )
         else
-          LastProjectFolderPersistenceHelper.update_last_project_folder(
+          Helper.update_last_project_folder(
             user:,
             project_folder:,
             origin_folder_id: project_storage.project_folder_id
